Move get_available_modes() function to confirmation.py
- Moved get_available_modes() from agent.py to confirmation.py for better organization
- Updated imports in agent.py to import the function from confirmation.py
- Updated test imports in test_mode_switching.py
- All tests pass and linting checks succeed
