+++ openhands_cli/acp_impl/confirmation.py
@@ -8,6 +8,7 @@ from acp import Client
 from acp.schema import (
     AllowedOutcome,
     PermissionOption,
+    SessionMode,
     ToolCallUpdate,
 )
 
@@ -71,6 +72,34 @@ PERMISSION_OPTIONS = [
         kind="allow_once",
     ),
 ]
+
+
+def get_available_modes() -> list[SessionMode]:
+    """Get list of available session modes.
+
+    Returns:
+        List of SessionMode objects representing available modes
+    """
+    return [
+        SessionMode(
+            id="always-ask",
+            name="Always Ask Mode",
+            description="Request permission before making any changes",
+        ),
+        SessionMode(
+            id="always-approve",
+            name="Always Approve Mode",
+            description="Automatically approve all actions without asking",
+        ),
+        SessionMode(
+            id="llm-approve",
+            name="LLM Approve Mode",
+            description=(
+                "Use LLM security analyzer to auto-approve safe actions. "
+                "Only ask for permission on high risk actions."
+            ),
+        ),
+    ]
 
 
 # Dispatch dictionary for handling user choices
